Fix map & refacto
