+++ MakgeolliGuru/app/src/main/java/com/example/makgeolliguru/InfoFragment.java
@@ -11,9 +11,9 @@ import android.widget.TextView;
 import androidx.fragment.app.Fragment;
 import androidx.fragment.app.FragmentManager;
 
-public class Info_Fragment extends Fragment {
+public class InfoFragment extends Fragment {
 
-    public Info_Fragment(){
+    public InfoFragment(){
         // require a empty public constructor
     }
 
@@ -37,7 +37,7 @@ public class Info_Fragment extends Fragment {
             public void onClick(View v) {
                 fragmentManager
                         .beginTransaction()
-                        .replace(R.id.flFragment, new Profile_Fragment())
+                        .replace(R.id.flFragment, new ProfileFragment())
                         .addToBackStack(null)  // Optional: Add to back stack for fragment navigation
                         .commit();
             }
