+++ MakgeolliGuru/app/src/main/java/com/example/makgeolliguru/MainActivity.java
@@ -1,117 +1,14 @@
 package com.example.makgeolliguru;
 
+import android.os.Bundle;
+import android.view.MenuItem;
 
 import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.fragment.app.Fragment;
 
-import android.content.SharedPreferences;
-import android.graphics.Bitmap;
-import android.graphics.BitmapFactory;
-import android.os.Bundle;
-import android.util.Log;
-import android.widget.Toast;
-
-import com.google.android.gms.maps.CameraUpdateFactory;
-import com.google.android.gms.maps.GoogleMap;
-import com.google.android.gms.maps.OnMapReadyCallback;
-import com.google.android.gms.maps.SupportMapFragment;
-import com.google.android.gms.maps.model.BitmapDescriptorFactory;
-import com.google.android.gms.maps.model.LatLng;
-import com.google.android.gms.maps.model.Marker;
-import com.google.android.gms.maps.model.MarkerOptions;
-
-/*public class MainActivity extends AppCompatActivity implements OnMapReadyCallback {
-
-    private GoogleMap myMap;
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_main);
-
-        SupportMapFragment mapFragment = (SupportMapFragment) getSupportFragmentManager().findFragmentById(R.id.map);
-        mapFragment.getMapAsync(this);
-    }
-
-
-    @Override
-    public void onMapReady(@NonNull GoogleMap googleMap) {
-        myMap = googleMap;
-
-        // Create Map center on South Korea
-        LatLng SouthKorea = new LatLng(37.566535, 126.9779692);
-        myMap.moveCamera(CameraUpdateFactory.newLatLng(SouthKorea));
-        myMap.getUiSettings().setZoomControlsEnabled(true);
-        myMap.getUiSettings().setMapToolbarEnabled(false);
-        //myMap.moveCamera(CameraUpdateFactory.zoomTo(10));
-        myMap.setInfoWindowAdapter(new CustomInfoWindowAdapter(MainActivity.this));
-
-        // Get data
-        //results = {"features":[{"type":"Feature","properties":{"mag":3,"place":"34km"},"geometry":{"type":"Point","coordinates":[127,35,8.7]} }]};
-
-        double[][] results = {{127, 35},{129.0756416 , 35.1795543 },{126.9779692, 37.566535},{128.9779692, 36.566535}};
-        String[][] infos = {{"127", "35", "2", "3", "3", "3","rice, corn", "Jeollanamdo", "14%", "???", "Yes", "Link" },{"129.0756416" , "35.1795543", "1", "1", "4", "3","rice", "Busan", "14%", "???", "No", "Link" },{"126.9779692", "37.566535", "2", "3", "3", "3","rice, nuts", "Seoul", "14%", "???", "Yes", "Link"},{"128.9779692", "36.566535", "2", "3", "3", "3","rice, corn", "Daegu", "14%", "???", "Yes", "Link"}};
-        // boucle for to display the map with markers
-        MarkerOptions marker;
-        for (int i = 0; i < results.length; i++) {
-
-            // Get the coordonate of each marker
-            double[] coords = results[i];
-            //double[] coords = {127, 35};
-            String[] info = infos[i];
-
-            LatLng latLng = new LatLng(coords[1], coords[0]);
-
-            /*String snippet = "<b>Sugar: </b>"+info[2]+ "\n"+
-                            "Acidity "+info[3]+ "\n"+
-                            "Texture "+info[4]+ "\n"+
-                            "Smell "+info[5]+ "\n"+
-                            "Flavor/ingredients "+info[6]+ "\n"+
-                            "Localisation "+info[7]+ "\n"+
-                            "% alcohol "+info[8]+ "\n"+
-                            "favorite food "+info[9]+ "\n"+
-                            "Artisanal "+info[10]+ "\n"+
-                            "<a href='"+info[11]+"' target='new'>link to kakao</a>"+ "\n"+"\n"+
-                            "<h3>Comment </h3>";*/
-
- /*           String snippet = String.format("Sugar: %s\n Acidity %s \nTexture %s\nSmell %s\nFlavor/ingredients %s\nLocalisation %s\n percent alcohol %s\nfavorite food %s\nArtisanal %s\n\n<h3>Comment </h3>", info[2], info[3], info[4], info[5], info[6], info[7], info[8], info[9], info[10]);
-
-            int height = 100;
-            int width = 100;
-            Bitmap b = BitmapFactory.decodeResource(getResources(), R.drawable.marker);
-            Bitmap smallMarker = Bitmap.createScaledBitmap(b, width, height, false);
-
-            marker = new MarkerOptions()
-                    .position(latLng)
-                    .title("Makgeolli name")
-                    .snippet(snippet)
-                    .icon(BitmapDescriptorFactory.fromBitmap(smallMarker));
-
-            Marker newMarker = myMap.addMarker(marker);
-            //newMarker.showInfoWindow();
-
-
-        }
-
-        // Set a listener for info window events.
-        myMap.setOnInfoWindowClickListener(this::onInfoWindowClick);
-
-    }
-
-
-    public void onInfoWindowClick(Marker marker) {
-        Toast.makeText(this, "Info window clicked",
-                Toast.LENGTH_SHORT).show();
-        myMap.setInfoWindowAdapter(new CustomInfoWindowAdapter(MainActivity.this));
-
-    }
-}*/
-
-
+import com.example.makgeolliguru.map.MapFragment;
 import com.google.android.material.bottomnavigation.BottomNavigationView;
-import android.view.MenuItem;
-
-import java.util.concurrent.ExecutionException;
 
 public class MainActivity extends AppCompatActivity implements BottomNavigationView.OnNavigationItemSelectedListener {
 
@@ -121,12 +18,12 @@ public class MainActivity extends AppCompatActivity implements BottomNavigationV
     public static final String FAVORITE_LIST = "com.example.makgeolliguru.FAVORITE_LIST";
     public static final String ARTICLE_LIST = "com.example.makgeolliguru.ARTICLE_LIST";
 
-
     // Bottom navigation behavior
     BottomNavigationView bottomNavigationView;
     MapFragment mapFragment = new MapFragment();
-    Learning_Fragment learningFragment = new Learning_Fragment();
-    Profile_Fragment profileFragment = new Profile_Fragment();
+    LearningFragment learningFragment = new LearningFragment();
+    ProfileFragment profileFragment = new ProfileFragment();
+
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         System.out.print(FAVORITE_LIST);
@@ -164,10 +61,8 @@ public class MainActivity extends AppCompatActivity implements BottomNavigationV
         }
         return false;
     }
-    public void setCurrentFragment(Fragment fragment){
+
+    public void setCurrentFragment(Fragment fragment) {
         getSupportFragmentManager().beginTransaction().replace(R.id.flFragment, fragment).commit();
     }
-
-
-
 }
