+++ MakgeolliGuru/app/src/main/java/com/example/makgeolliguru/map/CustomInfoWindowAdapter.java
@@ -1,4 +1,4 @@
-package com.example.makgeolliguru;
+package com.example.makgeolliguru.map;
 
 import android.content.Context;
 import android.view.LayoutInflater;
@@ -8,18 +8,15 @@ import android.widget.TextView;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
-import androidx.core.view.LayoutInflaterCompat;
-import androidx.core.widget.TextViewKt;
 
+import com.example.makgeolliguru.R;
 import com.google.android.gms.maps.GoogleMap;
 import com.google.android.gms.maps.model.Marker;
 
 public class CustomInfoWindowAdapter implements GoogleMap.InfoWindowAdapter {
     private final View mWindow;
-    private Context mContext;
 
     public CustomInfoWindowAdapter(Context context) {
-        mContext = context;
         mWindow = LayoutInflater.from(context).inflate(R.layout.custom_info_window, null);
     }
