+++ MakgeolliGuru/local.properties
@@ -1,12 +1,9 @@
-## This file is automatically generated by Android Studio.
-# Do not modify this file -- YOUR CHANGES WILL BE ERASED!
-#
-# This file should *NOT* be checked into Version Control Systems,
+## This file must *NOT* be checked into Version Control Systems,
 # as it contains information specific to your local configuration.
 #
 # Location of the SDK. This is only used by Gradle.
 # For customization when using a Version Control System, please read the
 # header note.
-sdk.dir=C\:\\Users\\julie\\AppData\\Local\\Android\\Sdk
-
+#Fri Nov 08 14:40:40 CET 2024
 MAPS_API_KEY="AIzaSyDes_qg_hfzD7Ak-BWFhCPO-VDboDglz2U"
+sdk.dir=/Users/chameyrac/Library/Android/sdk
